Clean out my TODO!!
So satisfying to open the TODO and find that most of it is already
implemented.

Added a new one, for configurable polling interval (which carries with
it warnings if it's too short.. which means a warn method.. which means
abstracting it's output..)

On the upside, it'll make unit testing similar.

*Scurries off to add that to TODO*